Split building of fluidsynth - this probably breaks other toolchains
Fluidsynth depends on glib which depends on ffi and intl.

I can't really get multi arch builds working for glib. So it has to be
built once for each arch. The other libraries does support multi arch
compilation so we need to break out the dependencies to separate build
stages. (They are also installed as separate libraries so it makes sense
to me anyway).