Add toolchain building iOS libraries for armv7, armv7s and arm64
It's almost possible to utilize the common apple toolchain. The
problematic libs are curl and glib.

This can be worked on maybe by tweaking the configure scripts:
1. Curl can't determine the size of size_t when building 32 bit archs
(armv7 and armv7s) and 64 bit arch. However reading the source it
looks like all arm-based CPUs are handled equally using 32 bit sizes.
So it's maybe possible to tweak the configure script to set the size in
multi arch builds correct.

2. Glib fail to find the poll headers and zlib in multi arch builds.
It's not 32/64 bit dependent either because trying to build for armv7
and armv7s fails with the same reason. I've been trying to understand
why but couldn't find an answer. It was faster to add a workaround by
splitting the build scripts for fluidsynth and re-build glib for each
arch and then create a fat lib by the produced artifacts.